Feat: 팔로잉, 팔로위 유저 정보 엔드포인트 구현

+++ src/main/java/project/closet/dto/response/FollowListResponse.java
@@ -8,7 +8,7 @@ public record FollowListResponse(
         String nextCursor,
         UUID nextIdAfter,
         boolean hasNext,
-        int totalCount,
+        long totalCount,
         String sortBy,
         String sortDirection
 ) {
